LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/css/css-fonts/math-script-level-and-math-style/math-style-001.tentative-expected.txt


FAIL Initial value of math-style assert_equals: expected "normal" but got ""
FAIL Specified values of math-style assert_equals: expected "compact" but got ""
FAIL Inherited values of math-style assert_equals: expected "compact" but got ""
